Pass TraktManager to Route instead of using injection context, add movie endpoints
I experimented with using a dependency injection context so I can set the TraktManager once and have access to the instance from Route without having to pass the TraktManager through each `Resource` class, but with Swift concurrency and tests I found this to be not too safe. You would expect that the `traktManager` instance you created a resource with would perform the request so I removed it. I am still working on the architecture for the resource classes. I was going to use static functions within a Resource for functions that didn't need a specific shared id for the requests, such as series id for shows, or a user id for User requests, but have moved to using a separate resource struct so I can pass in the TraktManager to those endpoints. It's better than having `traktManager.resource.endpoint(traktManager` or anything odd like that.