Refactored authentication into its own class for better testability and API customization
Since the first version of TraktKit authentication was handled automatically by storing the `access_token` and `refresh_token` in the keychain, and the token expiration date in UserDefaults. This was good because consumers of the API didn't have to really worry about handling that, but it also wasn't very testable, and if you wanted to use your own storage or a different keychain implementation, you couldn't without forking the repo and changing it. This commit changes that by extracting the authentication handling from `TraktManager` into its own protocol `TraktAuthentication`, with a concrete implementation that matches the existing behavior `KeychainTraktAuthentication`. There is even a mock implementation now `TraktMockAuthStorage` that can be used for testing without overwriting your credentials in the keychain. You will now need to call `TraktManager.refreshCurrentAuthState()`, or use the async initializer before making requests to ensure `TraktManager` has the latest `access_token` if making authenticated requests.