Remove custom request handling for `/users/*/watching`
The `/users/*/watching` endpoints returns data if a user is currently watching something, or 204 with no data if the user is not watching anything. Rather than having a custom request handler for this, I've updated the Swift model that is decoded from the data to handle no data being returned. Since both responses have a 2xx response, an error is not thrown unless the data could not be decoded.