Remove custom request handling for `/checkin`
Removed the custom request handling for `TraktCheckin`. Unfortunately because the `/checkin` endpoint returns an expiration date with the status code `409`, the response handling that happens before attempting to decode the data will throw an error. Since this appears to be the only endpoint that returns alternate data for a non-2xx response that I can think of, and because you can get the same data from `/users/*/watching`, I figured this will be ok. You can check if the error thrown is `TraktManager.TraktError.resourceAlreadyCreated` to make that request and get the end date of a checkin.